backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
psycopg2-binary
qrcode
Pillow
pytest
pytest-django
requests
drf-spectacular
